feat: long contest 3

+++ LongContest1/I.cpp
@@ -7,7 +7,7 @@ int main(){
     int Sz = problem.size();
     long long int sum = 0;
     for(int i = 0; i < Sz; i ++){
-        sum += (problem[i] - 'A' + 1) * pow(26, (Sz - i - 1));
+        sum += (problem[i] - 'A' + 1) * (long long)(pow(26, (Sz - i - 1)));
     }
-    cout << sum;
+    cout << (long long)sum;
 }
